refactor: flatten package to src
