+++ .gitignore
@@ -1,2 +1,2 @@
 __pycache__/
-boomi_mcp/__pycache__/
+src/__pycache__/
