+++ server.py
@@ -1,4 +1,4 @@
-from boomi_mcp.server import main
+from src.server import main
 
 if __name__ == "__main__":
     main()
